+++ druid/src/core.rs
@@ -858,18 +858,16 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
             notifications: parent_notifications,
             ..
         } = ctx;
-        let mut sentinal = VecDeque::new();
         let self_id = self.id();
         let mut inner_ctx = EventCtx {
             state,
-            notifications: &mut sentinal,
+            notifications: parent_notifications,
             widget_state: &mut self.state,
             is_handled: false,
             is_root: false,
         };
 
-        for _ in 0..notifications.len() {
-            let notification = notifications.pop_front().unwrap();
+        for notification in notifications.drain(..) {
             // skip notifications that were submitted by our child
             if notification.source() != self_id {
                 let event = Event::Notification(notification);
@@ -878,20 +876,14 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                     inner_ctx.is_handled = false;
                 } else if let Event::Notification(notification) = event {
                     // we will try again with the next parent
-                    parent_notifications.push_back(notification);
+                    inner_ctx.notifications.push_back(notification);
                 } else {
+                    // could be unchecked but we avoid unsafe in druid :shrug:
                     unreachable!()
                 }
             } else {
-                parent_notifications.push_back(notification);
+                inner_ctx.notifications.push_back(notification);
             }
-        }
-
-        if !inner_ctx.notifications.is_empty() {
-            warn!(
-                "A Notification was submitted while handling another \
-            notification; the submitted notification will be ignored."
-            );
         }
     }